Rename test file, reintroduce legacy full code, and move namespaces around a bit.
