+++ cpp/CMakeLists.txt
@@ -331,10 +331,13 @@ add_library(cuspatial SHARED
             src/interpolate/cubic_spline.cu
             src/io/shp/legacy/polygon_shapefile_reader.cpp
             src/io/shp/legacy/polygon_shapefile_reader.cu
-            src/io/soa/polygon_soa_reader.cu
+            src/io/soa/legacy/polygon_soa_reader.cu
             src/io/soa/point_soa_reader.cu
-            src/io/soa/uint32_soa_reader.cu
+            src/io/soa/int32_soa_reader.cu
             src/io/soa/timestamp_soa_reader.cu
+            src/io/soa/legacy/point_soa_reader.cu
+            src/io/soa/legacy/uint32_soa_reader.cu
+            src/io/soa/legacy/timestamp_soa_reader.cu
             src/query/legacy/spatial_window_points.cu
             src/spatial/legacy/point_in_polygon.cu
             src/spatial/legacy/haversine.cu
